Move NonClassParent Entity
Move NonClassParent entity into the nonClassEvent module. This entity doesn't
really make sense by itself. Typically it's used by the nonClassEvent module
so it should live in the folder with that module
